Refactor ldg iterator to ldg ref wrapper
