+++ src/celeritas/grid/GridIdFinder.hh
@@ -8,8 +8,7 @@
 
 #include "corecel/Assert.hh"
 #include "corecel/Macros.hh"
-#include "corecel/cont/Span.hh"
-#include "corecel/data/LdgIterator.hh"
+#include "corecel/cont/LdgSpan.hh"
 #include "corecel/math/Algorithms.hh"
 
 namespace celeritas
