+++ src/corecel/data/detail/CollectionImpl.hh
@@ -12,18 +12,17 @@
 #    include <vector>
 
 #    include "corecel/data/DeviceVector.hh"
+#else
+#    include "DisabledStorage.hh"
 #endif
 
 #include "corecel/Assert.hh"
-#include "corecel/OpaqueId.hh"
 #include "corecel/Types.hh"
+#include "corecel/cont/LdgSpan.hh"
 #include "corecel/cont/Span.hh"
 #include "corecel/data/Copier.hh"
-#include "corecel/data/LdgIterator.hh"
 #include "corecel/data/PinnedAllocator.hh"
 #include "corecel/sys/Device.hh"
-
-#include "DisabledStorage.hh"
 
 namespace celeritas
 {
