+++ src/geocel/VolumeView.hh
@@ -8,7 +8,7 @@
 
 #include "corecel/Assert.hh"
 #include "corecel/Macros.hh"
-#include "corecel/data/LdgIterator.hh"
+#include "corecel/data/LdgRefWrapper.hh"
 #include "geocel/Types.hh"
 
 #include "VolumeData.hh"
