+++ src/orange/univ/detail/InfixEvaluator.hh
@@ -7,8 +7,8 @@
 #pragma once
 
 #include "corecel/Assert.hh"
+#include "corecel/cont/LdgSpan.hh"
 #include "corecel/cont/Span.hh"
-#include "corecel/data/LdgIterator.hh"
 #include "orange/OrangeTypes.hh"
 
 namespace celeritas
